+++ tiny_dnn/core/kernels/global_avepool_op_internal.h
@@ -15,7 +15,7 @@ inline void global_avepool_op_internal(
   tensor_t &out_data,
   const core::global_avepool_params &params,
   const bool layer_parallelize) {
-  for_i(layer_parallelize, in_data.size(), [&](int sample) {
+  for_i(layer_parallelize, in_data.size(), [&](size_t sample) {
     const vec_t &in = in_data[sample];
     vec_t &out      = out_data[sample];
 
@@ -34,7 +34,7 @@ inline void global_avepool_grad_op_internal(
   const tensor_t &curr_delta,
   const core::global_avepool_params &params,
   const bool layer_parallelize) {
-  for_i(layer_parallelize, prev_delta.size(), [&](int sample) {
+  for_i(layer_parallelize, prev_delta.size(), [&](size_t sample) {
     vec_t &prev       = prev_delta[sample];
     const vec_t &curr = curr_delta[sample];
